requirements.txt
google-cloud-texttospeech==2.16.3
google-generativeai
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1